requirements.txt
discord.py==2.3.2
yt-dlp==2023.3.4  # Use yt-dlp instead of youtube-dl
spotipy==2.24.0
Pillow==8.3.2
requests==2.31.0
fastapi==0.95.0
uvicorn==0.22.0  # Compatible with the latest FastAPI
pynacl

